Fix test hypothesis (#4380)
* fix:test-hypothesis

---------
